Make the bouncycastle build targets more regular
Before this commit:

The "bouncycastle", "bouncycastle-ocsp" and "bouncycastle-bcpkix"
targets do different things on host and on target WRT to the package
the compiled source ends up in. On host the libraries are in
org.bouncycastle while on target they are in
com.android.org.bouncycastle.

There are also "bouncycastle-unbundled", "bouncycastle-ocsp-unbundled"
and "bouncycastle-pcpkix-unbundled" which produce libraries in
org.bouncycastle.

This commit makes "bouncycastle", "bouncycastle-ocsp" and
"bouncycastle-bcpkix" consistently use com.android.org.bouncycastle.
All users wanting the original package should use
"bouncycastle-unbundled" and "bouncycastle-pcpkix-unbundled" instead.

This commit (and those in the same topic) are in preparation for
changing over from using jarjar to using generated source code.

Bug: 111734251
Bug: 111055375
Test: Build only
Change-Id: I16e3464d7d6ad0fc138353cbfa3ab572318484c1