Reduce dependencies of device bouncycastle targets
This commit changes the device bouncycastle targets to depend
on core.current.stubs (the parts of Android core libraries
that contribute to the public SDK).

This is intended to make it clearer that bouncycastle does
not depend on Android internals, which allows us to mostly
ignore it for the modularization work. It still may contribute
towards the intra-core / core platform APIs but that will be
dealt with later.

Also: Only the unbundled targets have "host_supported: true"
now. This may not strictly be required but non-unbundled ones
are intended for ART use only.

Bug: 113148576
Test: treehugger build
Change-Id: Iad7e86632b238ee614b278c09a453565c7ae139b