Annotate repackaged API with @UnsupportedAppUsage
Adds @UnsupportedAppUsage annotations to the parts of bouncycastle that
are in the hidden API. Allows the bouncycastle entries to be removed
from the frameworks/base/config/hiddenapi-light-greylist.txt.

This was tested by making and then manually checking that the generated
out/target/common/obj/PACKAGING/hiddenapi-light-greylist.txt was the
same (after sorting) before and after this change.

Bug: 117818301
Test: see above
Change-Id: Icdc7c2cf07710d808c32e09caf71867efec4d775