Add missing default constructors
Metalava was accidentally including default constructors in the API
generated for CorePlatformApi and IntraCoreApi annotations. This change
adds the constructor and appropriate annotations to allow the metalava
bug to be fixed without changing the API.

Tested by building the API with and without the Metalava fix and
checking that they were the same.

Bug: 119859108
Test: see above
Change-Id: Ibb537c88aaa0db2867cfef590c86bb414a98ee69