Add new bouncycastle filegroup for use by IKE.
IKEv2 and EAP require the use of SHA1 and MD4 hash algorithms to be
standards compliant when supporting legacy authentication schemes.
Specifically, EAP-SIM and EAP-AKA require SHA1 for implementing the
FIPS 186-2 algorithm in order to compute the master session key.
EAP-MSCHAPv2 requires MD4 for hashing the user's password and hashing the
password-hash.

To avoid adding the entire bouncycastle-unbundled jar into the IKE library
(due to classpath size concerns), a new minimal filegroup is defined and
used instead. This ensures that the jar size is as small as possible, as
it will be loaded into the bootpath along with IKE.

Bug: 142345379
Test: atest FrameworksIkeTests
Test: atest FrameworksNetTests
Change-Id: Ib89d21a27f527c731eb9ddaa153705aaf7093f1d